@@ -2,7 +2,7 @@
 name = "yewprint-doc"
 version = "0.1.0"
 authors = ["Cecile Tonglet <cecile.tonglet@cecton.com>"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -10,11 +10,14 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-wasm-bindgen = "=0.2"
-yew = "0.18"
-web-sys = { version = "0.3", features = ["Window", "MediaQueryList"] }
+wasm-bindgen = "0.2"
+web-sys = { version = "0.3", features = ["Window", "MediaQueryList", "Event", "HtmlInputElement"] }
+gloo = "0.6"
+yew = "0.19"
+yew-router = "0.16"
+# yew = { git = "https://github.com/yewstack/yew", branch = "master" }
+# yew-router = { git = "https://github.com/yewstack/yew", branch = "master" }
 yewprint = { path = ".." }
-yew-router = "0.15"
 
 # `wee_alloc` is a tiny allocator for wasm that is only ~1K in code size
 # compared to the default allocator's ~10K. It is slower than the default
