@@ -1,11 +1,7 @@
 use yew::prelude::*;
 
-pub struct Checkbox {
-    props: Props,
-}
-
 #[derive(Clone, PartialEq, Properties)]
-pub struct Props {
+pub struct CheckboxProps {
     #[prop_or_default]
     pub disabled: bool,
     #[prop_or_default]
@@ -15,56 +11,35 @@ pub struct Props {
     #[prop_or_default]
     pub checked: bool,
     #[prop_or_default]
-    pub onchange: Callback<ChangeData>,
+    pub onchange: Callback<Event>,
     #[prop_or_default]
     pub label: yew::virtual_dom::VNode,
     #[prop_or_default]
     pub indeterminate_state: bool,
 }
 
-impl Component for Checkbox {
-    type Message = ();
-    type Properties = Props;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Self { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        html! {
-            <label
-                class=classes!(
-                    "bp3-control", "bp3-checkbox",
-                    self.props.disabled.then(|| "bp3-disabled"),
-                    self.props.inline.then(|| "bp3-inline"),
-                    self.props.large.then(|| "bp3-large")
-                )
+#[function_component(Checkbox)]
+pub fn checkbox(props: &CheckboxProps) -> Html {
+    html! {
+        <label
+            class={classes!(
+                "bp3-control", "bp3-checkbox",
+                props.disabled.then(|| "bp3-disabled"),
+                props.inline.then(|| "bp3-inline"),
+                props.large.then(|| "bp3-large")
+            )}
+        >
+            <input
+                type="checkbox"
+                checked={props.checked}
+                onchange={props.onchange.clone()}
+                disabled={props.disabled}
+            />
+            <span
+                class="bp3-control-indicator"
             >
-                <input
-                    type="checkbox"
-                    checked={self.props.checked}
-                    onchange={self.props.onchange.clone()}
-                    disabled=self.props.disabled
-                />
-                <span
-                    class="bp3-control-indicator"
-                >
-                </span>
-                {self.props.label.clone()}
-            </label>
-        }
+            </span>
+            {props.label.clone()}
+        </label>
     }
 }
