@@ -1,9 +1,11 @@
+use std::marker::PhantomData;
+
 use crate::{Icon, IconName};
+use web_sys::HtmlSelectElement;
 use yew::prelude::*;
 
 pub struct HtmlSelect<T: Clone + PartialEq + 'static> {
-    props: HtmlSelectProps<T>,
-    link: ComponentLink<Self>,
+    phantom: PhantomData<T>,
 }
 
 #[derive(Clone, PartialEq, Properties)]
@@ -30,51 +32,44 @@ pub struct HtmlSelectProps<T: Clone + PartialEq + 'static> {
 }
 
 impl<T: Clone + PartialEq + 'static> Component for HtmlSelect<T> {
-    type Message = ChangeData;
+    type Message = Event;
     type Properties = HtmlSelectProps<T>;
 
-    fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
-        Self { props, link }
+    fn create(_ctx: &Context<Self>) -> Self {
+        Self {
+            phantom: PhantomData,
+        }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
-        let i = if let ChangeData::Select(select) = msg {
+    fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
+        let i = if let Some(select) = msg.target_dyn_into::<HtmlSelectElement>() {
             select.selected_index()
         } else {
-            unreachable!("unexpected ChangeData variant: {:?}", msg);
+            unreachable!("unexpected Event: {:?}", msg);
         };
         if i >= 0 {
             let i = i as usize;
-            let variant = self.props.options[i].0.clone();
-            self.props.onchange.emit(variant);
+            let variant = ctx.props().options[i].0.clone();
+            ctx.props().onchange.emit(variant);
         }
         false
     }
 
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let option_children = self
-            .props
+    fn view(&self, ctx: &Context<Self>) -> Html {
+        let option_children = ctx
+            .props()
             .options
             .iter()
             .map(|(value, label)| {
-                let selected = self
-                    .props
+                let selected = ctx
+                    .props()
                     .value
                     .as_ref()
                     .map(|x| value == x)
                     .unwrap_or_default();
 
                 html! {
-                    <option selected=selected>
+                    <option selected={selected}>
                         {label}
                     </option>
                 }
@@ -83,24 +78,24 @@ impl<T: Clone + PartialEq + 'static> Component for HtmlSelect<T> {
 
         html! {
             <div
-                class=classes!(
+                class={classes!(
                     "bp3-html-select",
-                    self.props.minimal.then(|| "bp3-minimal"),
-                    self.props.large.then(|| "bp3-large"),
-                    self.props.fill.then(|| "bp3-fill"),
-                    self.props.disabled.then(|| "bp3-disabled"),
-                    self.props.class.clone(),
-                )
+                    ctx.props().minimal.then(|| "bp3-minimal"),
+                    ctx.props().large.then(|| "bp3-large"),
+                    ctx.props().fill.then(|| "bp3-fill"),
+                    ctx.props().disabled.then(|| "bp3-disabled"),
+                    ctx.props().class.clone(),
+                )}
             >
                 <select
-                    disabled=self.props.disabled
-                    onchange={self.link.callback(|x| x)}
-                    title={self.props.title.clone()}
+                    disabled={ctx.props().disabled}
+                    onchange={ctx.link().callback(|x| x)}
+                    title={ctx.props().title.clone()}
                     value={"".to_string()}
                 >
                     {option_children}
                 </select>
-                <Icon icon=IconName::DoubleCaretVertical/>
+                <Icon icon={IconName::DoubleCaretVertical}/>
             </div>
         }
     }
