@@ -1,10 +1,6 @@
 use crate::Intent;
 use yew::prelude::*;
 
-pub struct ProgressBar {
-    props: ProgressBarProps,
-}
-
 #[derive(Clone, PartialEq, Properties)]
 pub struct ProgressBarProps {
     #[prop_or_default]
@@ -19,48 +15,27 @@ pub struct ProgressBarProps {
     pub class: Classes,
 }
 
-impl Component for ProgressBar {
-    type Message = ();
-    type Properties = ProgressBarProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Self { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let width = if let Some(value) = self.props.value {
-            // NOTE: nightly, issue #44095 for f32::clamp
-            // let percent = ((1000. * value).ceil() / 10.).clamp(0.,100.);
-            let percent = ((1000. * value).ceil() / 10.).max(0.).min(100.);
-            format!("width: {}%;", percent)
-        } else {
-            "".into()
-        };
-        html! {
-            <div
-                class=classes!(
-                    "bp3-progress-bar",
-                    self.props.intent,
-                    (!self.props.animate).then(|| "bp3-no-animation"),
-                    (!self.props.stripes).then(|| "bp3-no-stripes"),
-                    self.props.class.clone(),
-                )
-            >
-                <div class=classes!("bp3-progress-meter") style={{width}}/>
-            </div>
-        }
+#[function_component(ProgressBar)]
+pub fn progress_bar(props: &ProgressBarProps) -> Html {
+    let width = if let Some(value) = props.value {
+        // NOTE: nightly, issue #44095 for f32::clamp
+        // let percent = ((1000. * value).ceil() / 10.).clamp(0.,100.);
+        let percent = ((1000. * value).ceil() / 10.).max(0.).min(100.);
+        format!("width: {}%;", percent)
+    } else {
+        "".into()
+    };
+    html! {
+        <div
+            class={classes!(
+                "bp3-progress-bar",
+                props.intent,
+                (!props.animate).then(|| "bp3-no-animation"),
+                (!props.stripes).then(|| "bp3-no-stripes"),
+                props.class.clone(),
+            )}
+        >
+            <div class={classes!("bp3-progress-meter")} style={{width}}/>
+        </div>
     }
 }
