@@ -1,9 +1,5 @@
 use yew::prelude::*;
 
-pub struct Radio {
-    props: RadioProps,
-}
-
 #[derive(Clone, PartialEq, Properties)]
 pub struct RadioProps {
     #[prop_or_default]
@@ -17,59 +13,38 @@ pub struct RadioProps {
     #[prop_or_default]
     pub name: Option<String>,
     #[prop_or_default]
-    pub onchange: Option<Callback<ChangeData>>,
+    pub onchange: Option<Callback<Event>>,
     #[prop_or_default]
     pub label: yew::virtual_dom::VNode,
     #[prop_or_default]
     pub value: Option<String>,
 }
 
-impl Component for Radio {
-    type Message = ();
-    type Properties = RadioProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Self { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        html! {
-            <label
-                class=classes!(
-                    "bp3-control",
-                    "bp3-radio",
-                    self.props.disabled.then(|| "bp3-disabled"),
-                    self.props.inline.then(|| "bp3-inline"),
-                    self.props.large.then(|| "bp3-large"),
-                )
+#[function_component(Radio)]
+pub fn radio(props: &RadioProps) -> Html {
+    html! {
+        <label
+            class={classes!(
+                "bp3-control",
+                "bp3-radio",
+                props.disabled.then(|| "bp3-disabled"),
+                props.inline.then(|| "bp3-inline"),
+                props.large.then(|| "bp3-large"),
+            )}
+        >
+            <input
+                type="radio"
+                onchange={props.onchange.clone().unwrap_or_default()}
+                disabled={props.disabled}
+                value={props.value.clone().unwrap_or_default()}
+                checked={props.checked.unwrap_or(false)}
+                name={props.name.clone().unwrap_or_default()}
+            />
+            <span
+                class={classes!("bp3-control-indicator")}
             >
-                <input
-                    type="radio"
-                    onchange={self.props.onchange.clone().unwrap_or_default()}
-                    disabled=self.props.disabled
-                    value={self.props.value.clone().unwrap_or_default()}
-                    checked=self.props.checked.unwrap_or(false)
-                    name={self.props.name.clone().unwrap_or_default()}
-                />
-                <span
-                    class=classes!("bp3-control-indicator")
-                >
-                </span>
-                {self.props.label.clone()}
-            </label>
-        }
+            </span>
+            {props.label.clone()}
+        </label>
     }
 }
