@@ -1,10 +1,5 @@
 use crate::{if_html, Icon, IconName, Intent, Text};
-use std::borrow::Cow;
 use yew::prelude::*;
-
-pub struct Tag {
-    props: TagProps,
-}
 
 #[derive(Clone, PartialEq, Properties)]
 pub struct TagProps {
@@ -36,81 +31,60 @@ pub struct TagProps {
     #[prop_or_default]
     pub round: bool,
     #[prop_or_default]
-    pub title: Option<Cow<'static, str>>,
+    pub title: Option<String>,
     #[prop_or_default]
     pub class: Classes,
     #[prop_or_default]
-    pub style: Option<Cow<'static, str>>,
+    pub style: Option<String>,
 }
 
-impl Component for Tag {
-    type Message = ();
-    type Properties = TagProps;
+#[function_component(Tag)]
+pub fn tag(props: &TagProps) -> Html {
+    let icon = if_html!(let Some(icon) = props.icon => <Icon icon={icon} />);
 
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Tag { props }
-    }
+    let right_icon =
+        if_html!(let Some(right_icon) = props.right_icon => <Icon icon={right_icon} />);
 
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let icon = if_html!(let Some(icon) = self.props.icon => <Icon icon=icon />);
-
-        let right_icon =
-            if_html!(let Some(right_icon) = self.props.right_icon => <Icon icon=right_icon />);
-
-        let remove_button = if_html! {
-            let Some(callback) = self.props.onremove.clone() =>
-            html!(
-                <button
-                    class=classes!("bp3-tag-remove")
-                    onclick={callback}
-                    tabindex={self.props.interactive.then(|| "0")}
-                >
-                    <Icon icon=IconName::SmallCross />
-                </button>
-            )
-        };
-
-        html! {
-            <span
-                class=classes!(
-                    "bp3-tag",
-                    self.props.intent,
-                    self.props.active.then(|| "bp3-active"),
-                    self.props.fill.then(|| "bp3-fill"),
-                    self.props.interactive.then(|| "bp3-interactive"),
-                    self.props.large.then(|| "bp3-large"),
-                    self.props.minimal.then(|| "bp3-minimal"),
-                    self.props.round.then(|| "bp3-round"),
-                    self.props.class.clone(),
-                )
-                style=self.props.style.clone()
-                onclick={self.props.onclick.clone()}
+    let remove_button = if_html! {
+        let Some(callback) = props.onremove.clone() =>
+        html!(
+            <button
+                class={classes!("bp3-tag-remove")}
+                onclick={callback}
+                tabindex={props.interactive.then(|| "0")}
             >
-                {icon}
-                <Text
-                    class=classes!("bp3-fill")
-                    ellipsize={!self.props.multiline}
-                    title=self.props.title.clone()
-                    inline=true
-                >
-                    {self.props.children.clone()}
-                </Text>
-                {right_icon}
-                {remove_button}
-            </span>
-        }
+                <Icon icon={IconName::SmallCross} />
+            </button>
+        )
+    };
+
+    html! {
+        <span
+            class={classes!(
+                "bp3-tag",
+                props.intent,
+                props.active.then(|| "bp3-active"),
+                props.fill.then(|| "bp3-fill"),
+                props.interactive.then(|| "bp3-interactive"),
+                props.large.then(|| "bp3-large"),
+                props.minimal.then(|| "bp3-minimal"),
+                props.round.then(|| "bp3-round"),
+                props.class.clone(),
+            )}
+            style={props.style.clone()}
+            onclick={props.onclick.clone()}
+        >
+            {icon}
+            <Text
+                class={classes!("bp3-fill")}
+                ellipsize={!props.multiline}
+                title={props.title.clone()}
+                inline=true
+            >
+                {props.children.clone()}
+            </Text>
+            {right_icon}
+            {remove_button}
+        </span>
     }
 }
