@@ -1,9 +1,4 @@
-use std::borrow::Cow;
 use yew::prelude::*;
-
-pub struct Text {
-    props: TextProps,
-}
 
 #[derive(Clone, PartialEq, Properties)]
 pub struct TextProps {
@@ -17,44 +12,23 @@ pub struct TextProps {
     #[prop_or_default]
     pub inline: bool,
     #[prop_or_default]
-    pub title: Option<Cow<'static, str>>,
+    pub title: Option<String>,
     #[prop_or_default]
-    pub style: Option<Cow<'static, str>>,
+    pub style: Option<String>,
 }
 
-impl Component for Text {
-    type Message = ();
-    type Properties = TextProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Text { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        html! {
-            <@{if self.props.inline { "span" } else { "div"}}
-                class=classes!(
-                    self.props.class.clone(),
-                    self.props.ellipsize.then (|| "bp3-text-overflow-ellipsis"),
-                )
-                style=self.props.style.clone()
-                title=self.props.title.clone()
-            >
-                {self.props.children.clone()}
-            </@>
-        }
+#[function_component(Text)]
+pub fn text(props: &TextProps) -> Html {
+    html! {
+        <@{if props.inline { "span" } else { "div"}}
+            class={classes!(
+                props.class.clone(),
+                props.ellipsize.then (|| "bp3-text-overflow-ellipsis"),
+            )}
+            style={props.style.clone()}
+            title={props.title.clone()}
+        >
+            {props.children.clone()}
+        </@>
     }
 }
