@@ -1,10 +1,6 @@
 use crate::Intent;
 use yew::prelude::*;
 
-pub struct TextArea {
-    props: TextAreaProps,
-}
-
 #[derive(Clone, PartialEq, Properties)]
 pub struct TextAreaProps {
     #[prop_or_default]
@@ -22,44 +18,23 @@ pub struct TextAreaProps {
     #[prop_or_default]
     pub small: bool,
     #[prop_or_default]
-    pub onchange: Option<Callback<ChangeData>>,
+    pub onchange: Option<Callback<Event>>,
 }
 
-impl Component for TextArea {
-    type Message = ();
-    type Properties = TextAreaProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        TextArea { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let classes = classes!(
-            "bp3-input",
-            self.props.intent,
-            self.props.class.clone(),
-            self.props.fill.then(|| "bp3-fill"),
-            self.props.small.then(|| "bp3-small"),
-            self.props.large.then(|| "bp3-large"),
-        );
-        html! {
-            <textarea
-                class=classes
-                onchange=self.props.onchange.clone()
-            />
-        }
+#[function_component(TextArea)]
+pub fn text_area(props: &TextAreaProps) -> Html {
+    let classes = classes!(
+        "bp3-input",
+        props.intent,
+        props.class.clone(),
+        props.fill.then(|| "bp3-fill"),
+        props.small.then(|| "bp3-small"),
+        props.large.then(|| "bp3-large"),
+    );
+    html! {
+        <textarea
+            class={classes}
+            onchange={props.onchange.clone()}
+        />
     }
 }
