@@ -6,6 +6,7 @@ use std::cell::{Ref, RefCell, RefMut};
 use std::collections::hash_map::DefaultHasher;
 use std::collections::HashMap;
 use std::hash::{Hash, Hasher};
+use std::marker::PhantomData;
 use std::rc::Rc;
 use yew::prelude::*;
 
@@ -41,13 +42,13 @@ impl<T> From<id_tree::Tree<NodeData<T>>> for TreeData<T> {
     }
 }
 
-pub struct Tree<T: Clone> {
-    props: TreeProps<T>,
+pub struct Tree<T: Clone + PartialEq> {
     previous_expanded_state: RefCell<HashMap<u64, bool>>,
+    phantom: PhantomData<T>,
 }
 
 #[derive(Clone, PartialEq, Properties)]
-pub struct TreeProps<T: Clone> {
+pub struct TreeProps<T: Clone + PartialEq> {
     #[prop_or_default]
     pub is_expanded: bool,
     pub tree: TreeData<T>,
@@ -112,41 +113,32 @@ impl<T: Clone + PartialEq + 'static> Component for Tree<T> {
     type Message = ();
     type Properties = TreeProps<T>;
 
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Tree {
-            props,
+    fn create(_ctx: &Context<Self>) -> Self {
+        Self {
             previous_expanded_state: Default::default(),
+            phantom: PhantomData,
         }
     }
 
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, _msg: Self::Message) -> bool {
         true
     }
 
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let tree = self.props.tree.borrow();
+    fn view(&self, ctx: &Context<Self>) -> Html {
+        let tree = ctx.props().tree.borrow();
 
         let nodes = if let Some(root_id) = tree.root_node_id() {
-            self.render_children(root_id, 0)
+            self.render_children(ctx, root_id, 0)
         } else {
             html!()
         };
 
         html! {
-            <div class=classes!(
+            <div class={classes!(
                 "bp3-tree",
-                self.props.class.clone(),
-            )>
-                <ul class=classes!("bp3-tree-node-list")>
+                ctx.props().class.clone(),
+            )}>
+                <ul class={classes!("bp3-tree-node-list")}>
                     {nodes}
                 </ul>
             </div>
@@ -154,9 +146,15 @@ impl<T: Clone + PartialEq + 'static> Component for Tree<T> {
     }
 }
 
-impl<T: Clone> Tree<T> {
-    fn render_children(&self, node_id: &NodeId, depth: u32) -> yew::virtual_dom::VNode {
-        let tree = self.props.tree.borrow();
+// FIXME: The 'static bound here is probably wrong. Fix this at the end of PR.
+impl<T: 'static + Clone + PartialEq> Tree<T> {
+    fn render_children(
+        &self,
+        ctx: &Context<Self>,
+        node_id: &NodeId,
+        depth: u32,
+    ) -> yew::virtual_dom::VNode {
+        let tree = ctx.props().tree.borrow();
         let node = tree.get(node_id).unwrap();
         let children = node.children();
 
@@ -177,26 +175,26 @@ impl<T: Clone> Tree<T> {
                 let inner_nodes = if !data.is_expanded && !previous_is_expanded.unwrap_or(true) {
                     Default::default()
                 } else {
-                    self.render_children(node_id, depth + 1)
+                    self.render_children(ctx, node_id, depth + 1)
                 };
 
                 html! {
                     <TreeNode
-                        disabled=data.disabled
-                        has_caret=data.has_caret
-                        icon=data.icon
-                        icon_color=data.icon_color.clone()
-                        icon_intent=data.icon_intent
-                        is_expanded=data.is_expanded
-                        is_selected=data.is_selected
-                        label=data.label.clone()
-                        secondary_label=data.secondary_label.clone()
-                        on_collapse=self.props.on_collapse.clone()
-                        on_expand=self.props.on_expand.clone()
-                        onclick=self.props.onclick.clone()
-                        depth=depth
-                        node_id=node_id.clone()
-                        key=key
+                        disabled={data.disabled}
+                        has_caret={data.has_caret}
+                        icon={data.icon}
+                        icon_color={data.icon_color.clone()}
+                        icon_intent={data.icon_intent}
+                        is_expanded={data.is_expanded}
+                        is_selected={data.is_selected}
+                        label={data.label.clone()}
+                        secondary_label={data.secondary_label.clone()}
+                        on_collapse={ctx.props().on_collapse.clone()}
+                        on_expand={ctx.props().on_expand.clone()}
+                        onclick={ctx.props().onclick.clone()}
+                        depth={depth}
+                        node_id={node_id.clone()}
+                        key={key}
                     >
                         {inner_nodes}
                     </TreeNode>
@@ -207,7 +205,6 @@ impl<T: Clone> Tree<T> {
 }
 
 struct TreeNode {
-    props: TreeNodeProps,
     handler_caret_click: Callback<MouseEvent>,
     handler_click: Callback<MouseEvent>,
 }
@@ -260,34 +257,33 @@ impl Component for TreeNode {
     type Message = TreeNodeMessage;
     type Properties = TreeNodeProps;
 
-    fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         TreeNode {
-            handler_caret_click: link.callback(TreeNodeMessage::CaretClick),
-            handler_click: link.callback(TreeNodeMessage::Click),
-            props,
+            handler_caret_click: ctx.link().callback(TreeNodeMessage::CaretClick),
+            handler_click: ctx.link().callback(TreeNodeMessage::Click),
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
-        if self.props.disabled {
+    fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
+        if ctx.props().disabled {
             return false;
         }
 
         match msg {
             TreeNodeMessage::CaretClick(event) => {
-                if self.props.is_expanded {
-                    if let Some(on_collapse) = self.props.on_collapse.as_ref() {
+                if ctx.props().is_expanded {
+                    if let Some(on_collapse) = ctx.props().on_collapse.as_ref() {
                         event.stop_propagation();
-                        on_collapse.emit((self.props.node_id.clone(), event));
+                        on_collapse.emit((ctx.props().node_id.clone(), event));
                     }
-                } else if let Some(on_expand) = self.props.on_expand.as_ref() {
+                } else if let Some(on_expand) = ctx.props().on_expand.as_ref() {
                     event.stop_propagation();
-                    on_expand.emit((self.props.node_id.clone(), event));
+                    on_expand.emit((ctx.props().node_id.clone(), event));
                 }
             }
             TreeNodeMessage::Click(event) => {
-                if let Some(onclick) = self.props.onclick.as_ref() {
-                    onclick.emit((self.props.node_id.clone(), event));
+                if let Some(onclick) = ctx.props().onclick.as_ref() {
+                    onclick.emit((ctx.props().node_id.clone(), event));
                 }
             }
         }
@@ -295,38 +291,23 @@ impl Component for TreeNode {
         false
     }
 
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            // crate::log!(
-            //     "rerender {:?} {} {:?}",
-            //     self.props.node_id,
-            //     self.props.children == props.children,
-            //     self.props.icon,
-            // );
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        let content_style = format!("padding-left: {}px;", 23 * self.props.depth);
+    fn view(&self, ctx: &Context<Self>) -> Html {
+        let content_style = format!("padding-left: {}px;", 23 * ctx.props().depth);
 
         html! {
-            <li class=classes!(
+            <li class={classes!(
                 "bp3-tree-node",
-                self.props.is_selected.then(|| "bp3-tree-node-selected")
-            )>
+                ctx.props().is_selected.then(|| "bp3-tree-node-selected")
+            )}>
                 <div
                     class="bp3-tree-node-content"
-                    style=content_style
-                    onclick=self.handler_click.clone()
+                    style={content_style}
+                    onclick={self.handler_click.clone()}
                 >
                     {
-                        if self.props.has_caret {
+                        if ctx.props().has_caret {
                             let mut class = Classes::from("bp3-tree-node-caret");
-                            class.push(if self.props.is_expanded {
+                            class.push(if ctx.props().is_expanded {
                                 "bp3-tree-node-caret-open"
                             } else {
                                 "bp3-tree-node-caret-closed"
@@ -334,9 +315,9 @@ impl Component for TreeNode {
 
                             html! {
                                 <Icon
-                                    class=classes!(class.to_string())
-                                    icon=IconName::ChevronRight
-                                    onclick=self.handler_caret_click.clone()
+                                    class={classes!(class.to_string())}
+                                    icon={IconName::ChevronRight}
+                                    onclick={self.handler_caret_click.clone()}
                                 />
                             }
                         } else {
@@ -346,16 +327,16 @@ impl Component for TreeNode {
                         }
                     }
                     <Icon
-                        class=classes!("bp3-tree-node-icon")
-                        icon=self.props.icon.unwrap_or_default()
-                        color=self.props.icon_color.clone()
-                        intent=self.props.icon_intent
+                        class={classes!("bp3-tree-node-icon")}
+                        icon={ctx.props().icon.unwrap_or_default()}
+                        color={ctx.props().icon_color.clone()}
+                        intent={ctx.props().icon_intent}
                     />
-                    <span class=classes!("bp3-tree-node-label")>{self.props.label.clone()}</span>
+                    <span class={classes!("bp3-tree-node-label")}>{ctx.props().label.clone()}</span>
                     {
-                        if let Some(label) = self.props.secondary_label.clone() {
+                        if let Some(label) = ctx.props().secondary_label.clone() {
                             html!(
-                                <span class=classes!("bp3-tree-node-secondary-label")>
+                                <span class={classes!("bp3-tree-node-secondary-label")}>
                                     {label}
                                 </span>
                             )
@@ -364,9 +345,9 @@ impl Component for TreeNode {
                         }
                     }
                 </div>
-                <Collapse is_open=self.props.is_expanded>
-                    <ul class=classes!("bp3-tree-node-list")>
-                        {self.props.children.clone()}
+                <Collapse is_open={ctx.props().is_expanded}>
+                    <ul class={classes!("bp3-tree-node-list")}>
+                        {ctx.props().children.clone()}
                     </ul>
                 </Collapse>
             </li>
