@@ -22,19 +22,21 @@ use crate::tag::*;
 use crate::text::*;
 use crate::text_area::*;
 use crate::tree::*;
-use std::borrow::Cow;
 use yew::prelude::*;
-use yew_router::{
-    agent::{RouteAgentDispatcher, RouteRequest},
-    router::Router,
-    Switch,
-};
+use yew_router::prelude::*;
 use yewprint::{IconName, Menu, MenuItem};
 
+#[function_component(AppRoot)]
+pub fn app_root() -> Html {
+    html! {
+        <BrowserRouter>
+            <App></App>
+        </BrowserRouter>
+    }
+}
+
 pub struct App {
-    link: ComponentLink<Self>,
     dark_theme: bool,
-    route_dispatcher: RouteAgentDispatcher,
 }
 
 pub enum Msg {
@@ -46,33 +48,30 @@ impl Component for App {
     type Message = Msg;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(_ctx: &Context<Self>) -> Self {
         App {
             dark_theme: web_sys::window()
                 .and_then(|x| x.match_media("(prefers-color-scheme: dark)").ok().flatten())
                 .map(|x| x.matches())
                 .unwrap_or(true),
-            link,
-            route_dispatcher: RouteAgentDispatcher::new(),
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
         match msg {
             Msg::ToggleLight => self.dark_theme ^= true,
             Msg::GoToMenu(doc_menu) => {
-                self.route_dispatcher
-                    .send(RouteRequest::ChangeRoute(doc_menu.into()));
+                if let Some(history) = ctx.link().history() {
+                    history.push(doc_menu);
+                } else {
+                    gloo::console::warn!("Could not get history from Context");
+                }
             }
         }
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         let netlify_badge = if self.dark_theme {
             "https://www.netlify.com/img/global/badges/netlify-color-accent.svg"
         } else {
@@ -89,223 +88,178 @@ impl Component for App {
             IconName::Moon
         };
 
-        html! {
-            <div class=classes!("docs-root", self.dark_theme.then(|| "bp3-dark"))>
-                <div class=classes!("docs-app")>
-                    <div class=classes!("docs-nav-wrapper")>
-                        <div class=classes!("docs-nav")>
-                            <div class=classes!("docs-nav-title")>
-                                <a class=classes!("docs-logo") href="/">
-                                    {crate::include_raw_html!("logo.svg")}
-                                </a>
-                                <div>
-                                    <div class=classes!("bp3-navbar-heading", "docs-heading")>
-                                        {"Yewprint"}
-                                    </div>
-                                    <a
-                                        class=classes!("bp3-text-muted")
-                                        href="https://github.com/yewprint/yewprint"
-                                        target="_blank"
-                                    >
-                                        <small>{"View on GitHub"}</small>
-                                    </a>
-                                </div>
+        let menu = html! {
+            <Menu>
+                <MenuItem
+                    text={html!(go_to_theme_label)}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::ToggleLight)}
+                    icon={go_to_theme_icon}
+                />
+                <MenuItem
+                    text={html!("Button")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Button))}
+                />
+                <MenuItem
+                    text={html!("ButtonGroup")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::ButtonGroup))}
+                />
+                <MenuItem
+                    text={html!("Callout")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Callout))}
+                />
+                <MenuItem
+                    text={html!("Card")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Card))}
+                />
+                <MenuItem
+                    text={html!("Checkbox")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Checkbox))}
+                />
+                <MenuItem
+                    text={html!("Collapse")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Collapse))}
+                />
+                <MenuItem
+                    text={html!("ControlGroup")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::ControlGroup))}
+                    />
+                <MenuItem
+                    text={html!("Divider")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Divider))}
+                />
+                <MenuItem
+                    text={html!("HtmlSelect")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::HtmlSelect))}
+                />
+                <MenuItem
+                    text={html!("Icon")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Icon))}
+                />
+                <MenuItem
+                    text={html!("InputGroup")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::InputGroup))}
+                />
+                <MenuItem
+                    text={html!("Menu")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
+                />
+                <MenuItem
+                    text={html!("NumericInput")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::NumericInput))}
+                />
+                <MenuItem
+                    text={html!("PanelStack")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::PanelStack))}
+                />
+                <MenuItem
+                    text={html!("ProgressBar")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::ProgressBar))}
+                />
+                <MenuItem
+                    text={html!("Radio")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Radio))}
+                />
+                <MenuItem
+                    text={html!("Slider")}
+                    onclick={ctx.link().callback(|_| Msg::GoToMenu(DocMenu::Slider))}
+                />
+                <MenuItem
+                    text={html!("Spinner")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Spinner))}
+                />
+                <MenuItem
+                    text={html!("Switch")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Switch))}
+                />
+                <MenuItem
+                    text={html!("Tabs")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Tabs))}
+                />
+                <MenuItem
+                    text={html!("Tag")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Tag))}
+                />
+                <MenuItem
+                    text={html!("Text")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Text))}
+                />
+                <MenuItem
+                    text={html!("TextArea")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::TextArea))}
+                />
+                <MenuItem
+                    text={html!("Tree")}
+                    onclick={ctx.link()
+                        .callback(|_| Msg::GoToMenu(DocMenu::Tree))}
+                />
+                // NOTE: thanks to keep this list of <MenuItem> sorted
+                //       alphabetically (except for the light switch)
+            </Menu>
+        };
+
+        let navigation = html! {
+            <div class={classes!("docs-nav-wrapper")}>
+                <div class={classes!("docs-nav")}>
+                    <div class={classes!("docs-nav-title")}>
+                        <a class={classes!("docs-logo")} href="/">
+                            {crate::include_raw_html!("logo.svg")}
+                        </a>
+                        <div>
+                            <div class={classes!("bp3-navbar-heading", "docs-heading")}>
+                                {"Yewprint"}
                             </div>
-                            <Menu>
-                                <MenuItem
-                                    text={html!(go_to_theme_label)}
-                                    onclick=self.link
-                                        .callback(|_| Msg::ToggleLight)
-                                    icon=go_to_theme_icon
-                                />
-                                <MenuItem
-                                    text={html!("Button")}
-                                    href=Cow::Borrowed("#button")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Button))
-                                />
-                                <MenuItem
-                                    text={html!("ButtonGroup")}
-                                    href=Cow::Borrowed("#button-group")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::ButtonGroup))
-                                />
-                                <MenuItem
-                                    text={html!("Callout")}
-                                    href=Cow::Borrowed("#callout")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Callout))
-                                />
-                                <MenuItem
-                                    text={html!("Card")}
-                                    href=Cow::Borrowed("#card")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Card))
-                                />
-                                <MenuItem
-                                    text={html!("Checkbox")}
-                                    href=Cow::Borrowed("#checkbox")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Checkbox))
-                                />
-                                <MenuItem
-                                    text={html!("Collapse")}
-                                    href=Cow::Borrowed("#collapse")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Collapse))
-                                />
-                                <MenuItem
-                                    text={html!("ControlGroup")}
-                                    href=Cow::Borrowed("#control-group")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::ControlGroup))
-                                    />
-                                <MenuItem
-                                    text={html!("Divider")}
-                                    href=Cow::Borrowed("#divider")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Divider))
-                                />
-                                <MenuItem
-                                    text={html!("HtmlSelect")}
-                                    href=Cow::Borrowed("#html-select")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::HtmlSelect))
-                                />
-                                <MenuItem
-                                    text={html!("Icon")}
-                                    href=Cow::Borrowed("#icon")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Icon))
-                                />
-                                <MenuItem
-                                    text={html!("InputGroup")}
-                                    href=Cow::Borrowed("#input-group")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::InputGroup))
-                                />
-                                <MenuItem
-                                    text={html!("Menu")}
-                                    href=Cow::Borrowed("#menu")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Menu))
-                                />
-                                <MenuItem
-                                    text={html!("NumericInput")}
-                                    href=Cow::Borrowed("#numeric-input")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::NumericInput))
-                                />
-                                <MenuItem
-                                    text={html!("PanelStack")}
-                                    href=Cow::Borrowed("#panel-stack")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::PanelStack))
-                                />
-                                <MenuItem
-                                    text={html!("ProgressBar")}
-                                    href=Cow::Borrowed("#progress-bar")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::ProgressBar))
-                                />
-                                <MenuItem
-                                    text={html!("Radio")}
-                                    href=Cow::Borrowed("#radio")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Radio))
-                                />
-                                <MenuItem
-                                    text={html!("Slider")}
-                                    href=Cow::Borrowed("#slider")
-                                    onclick=self.link.callback(|_| Msg::GoToMenu(DocMenu::Slider))
-                                />
-                                <MenuItem
-                                    text={html!("Spinner")}
-                                    href=Cow::Borrowed("#spinner")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Spinner))
-                                />
-                                <MenuItem
-                                    text={html!("Switch")}
-                                    href=Cow::Borrowed("#switch")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Switch))
-                                />
-                                <MenuItem
-                                    text={html!("Tabs")}
-                                    href=Cow::Borrowed("#tabs")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Tabs))
-                                />
-                                <MenuItem
-                                    text={html!("Tag")}
-                                    href=Cow::Borrowed("#tag")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Tag))
-                                />
-                                <MenuItem
-                                    text={html!("Text")}
-                                    href=Cow::Borrowed("#text")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Text))
-                                />
-                                <MenuItem
-                                    text={html!("TextArea")}
-                                    href=Cow::Borrowed("#textarea")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::TextArea))
-                                />
-                                <MenuItem
-                                    text={html!("Tree")}
-                                    href=Cow::Borrowed("#tree")
-                                    onclick=self.link
-                                        .callback(|_| Msg::GoToMenu(DocMenu::Tree))
-                                />
-                                // NOTE: thanks to keep this list of <MenuItem> sorted
-                                //       alphabetically (except for the light switch)
-                            </Menu>
-                            <div class="docs-nav-sponsors">
-                                <a href=Cow::Borrowed("https://www.netlify.com")>
-                                    <img
-                                        src=netlify_badge
-                                        alt="Deploys by Netlify"
-                                    />
-                                </a>
-                            </div>
+                            <a
+                                class={classes!("bp3-text-muted")}
+                                href="https://github.com/yewprint/yewprint"
+                                target="_blank"
+                            >
+                                <small>{"View on GitHub"}</small>
+                            </a>
                         </div>
                     </div>
-                    <main class=classes!("docs-content-wrapper") role="main">
-                        <div class=classes!("docs-page")>
-                            <Router<DocMenu, ()>
-                                render=Router::render(|switch: DocMenu| {
-                                    match switch {
-                                        DocMenu::Button | DocMenu::Home => html! (<ButtonDoc />),
-                                        DocMenu::ButtonGroup => html! (<ButtonGroupDoc />),
-                                        DocMenu::Callout => html!(<CalloutDoc />),
-                                        DocMenu::Card => html!(<CardDoc />),
-                                        DocMenu::Checkbox => html!(<CheckboxDoc />),
-                                        DocMenu::Collapse => html!(<CollapseDoc />),
-                                        DocMenu::ControlGroup => html!(<ControlGroupDoc />),
-                                        DocMenu::Divider => html!(<DividerDoc />),
-                                        DocMenu::HtmlSelect => html!(<HtmlSelectDoc />),
-                                        DocMenu::Icon => html!(<IconDoc />),
-                                        DocMenu::InputGroup => html!(<InputGroupDoc />),
-                                        DocMenu::Menu => html!(<MenuDoc />),
-                                        DocMenu::NumericInput => html!(<NumericInputDoc />),
-                                        DocMenu::PanelStack => html!(<PanelStackDoc />),
-                                        DocMenu::ProgressBar => html!(<ProgressBarDoc />),
-                                        DocMenu::Radio => html!(<RadioDoc />),
-                                        DocMenu::Slider => html!(<SliderDoc />),
-                                        DocMenu::Spinner => html!(<SpinnerDoc />),
-                                        DocMenu::Switch => html!(<SwitchDoc />),
-                                        DocMenu::Tabs => html!(<TabsDoc />),
-                                        DocMenu::Tag => html!(<TagDoc />),
-                                        DocMenu::Text => html!(<TextDoc />),
-                                        DocMenu::TextArea => html!(<TextAreaDoc />),
-                                        DocMenu::Tree => html!(<TreeDoc />),
-                                    }
-                                })
+                    {{ menu }}
+                    <div class="docs-nav-sponsors">
+                        <a href={"https://www.netlify.com"}>
+                            <img
+                                src={netlify_badge}
+                                alt="Deploys by Netlify"
                             />
+                        </a>
+                    </div>
+                </div>
+            </div>
+        };
+
+        html! {
+            <div class={classes!("docs-root", self.dark_theme.then(|| "bp3-dark"))}>
+                <div class={classes!("docs-app")}>
+                    {{ navigation }}
+                    <main class={classes!("docs-content-wrapper")} role="main">
+                        <div class={classes!("docs-page")}>
+                                <Switch<DocMenu> render={Switch::render(switch)} />
                         </div>
                     </main>
                 </div>
@@ -314,56 +268,85 @@ impl Component for App {
     }
 }
 
-#[derive(Debug, Copy, Clone, Switch)]
+fn switch(route: &DocMenu) -> Html {
+    match route {
+        DocMenu::Button | DocMenu::Home => html! (<ButtonDoc />),
+        DocMenu::ButtonGroup => html! (<ButtonGroupDoc />),
+        DocMenu::Callout => html!(<CalloutDoc />),
+        DocMenu::Card => html!(<CardDoc />),
+        DocMenu::Checkbox => html!(<CheckboxDoc />),
+        DocMenu::Collapse => html!(<CollapseDoc />),
+        DocMenu::ControlGroup => html!(<ControlGroupDoc />),
+        DocMenu::Divider => html!(<DividerDoc />),
+        DocMenu::HtmlSelect => html!(<HtmlSelectDoc />),
+        DocMenu::Icon => html!(<IconDoc />),
+        DocMenu::InputGroup => html!(<InputGroupDoc />),
+        DocMenu::Menu => html!(<MenuDoc />),
+        DocMenu::NumericInput => html!(<NumericInputDoc />),
+        DocMenu::PanelStack => html!(<PanelStackDoc />),
+        DocMenu::ProgressBar => html!(<ProgressBarDoc />),
+        DocMenu::Radio => html!(<RadioDoc />),
+        DocMenu::Slider => html!(<SliderDoc />),
+        DocMenu::Spinner => html!(<SpinnerDoc />),
+        DocMenu::Switch => html!(<SwitchDoc />),
+        DocMenu::Tabs => html!(<TabsDoc />),
+        DocMenu::Tag => html!(<TagDoc />),
+        DocMenu::Text => html!(<TextDoc />),
+        DocMenu::TextArea => html!(<TextAreaDoc />),
+        DocMenu::Tree => html!(<TreeDoc />),
+    }
+}
+
+#[derive(PartialEq, Clone, Routable)]
 pub enum DocMenu {
-    #[to = "/#button-group"]
+    #[at("/button-group")]
     ButtonGroup,
-    #[to = "/#button"]
+    #[at("/button")]
     Button,
-    #[to = "/#callout"]
+    #[at("/callout")]
     Callout,
-    #[to = "/#card"]
+    #[at("/card")]
     Card,
-    #[to = "/#checkbox"]
+    #[at("/checkbox")]
     Checkbox,
-    #[to = "/#collapse"]
+    #[at("/collapse")]
     Collapse,
-    #[to = "/#control-group"]
+    #[at("/control-group")]
     ControlGroup,
-    #[to = "/#html-select"]
+    #[at("/html-select")]
     HtmlSelect,
-    #[to = "/#divider"]
+    #[at("/divider")]
     Divider,
-    #[to = "/#icon"]
+    #[at("/icon")]
     Icon,
-    #[to = "/#input-group"]
+    #[at("/input-group")]
     InputGroup,
-    #[to = "/#menu"]
+    #[at("/menu")]
     Menu,
-    #[to = "/#numeric-input"]
+    #[at("/numeric-input")]
     NumericInput,
-    #[to = "/#panel-stack"]
+    #[at("/panel-stack")]
     PanelStack,
-    #[to = "/#progress-bar"]
+    #[at("/progress-bar")]
     ProgressBar,
-    #[to = "/#radio"]
+    #[at("/radio")]
     Radio,
-    #[to = "/#slider"]
+    #[at("/slider")]
     Slider,
-    #[to = "/#spinner"]
+    #[at("/spinner")]
     Spinner,
-    #[to = "/#switch"]
+    #[at("/switch")]
     Switch,
-    #[to = "/#tabs"]
+    #[at("/tabs")]
     Tabs,
-    #[to = "/#tag"]
+    #[at("/tag")]
     Tag,
-    #[to = "/#textarea"]
+    #[at("/textarea")]
     TextArea,
-    #[to = "/#text"]
+    #[at("/text")]
     Text,
-    #[to = "/#tree"]
+    #[at("/tree")]
     Tree,
-    #[to = "/"]
+    #[at("/")]
     Home,
 }
