@@ -1,10 +1,5 @@
-use std::borrow::Cow;
 use yew::prelude::*;
 use yewprint::{Button, ButtonGroup, IconName};
-
-pub struct Example {
-    props: ExampleProps,
-}
 
 #[derive(Clone, PartialEq, Properties)]
 pub struct ExampleProps {
@@ -14,40 +9,19 @@ pub struct ExampleProps {
     pub vertical: bool,
 }
 
-impl Component for Example {
-    type Message = ();
-    type Properties = ExampleProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Example { props }
-    }
-
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
-        html! {
-            <ButtonGroup
-                minimal=self.props.minimal
-                fill=self.props.fill
-                large=self.props.large
-                vertical=self.props.vertical
-                style=Cow::Borrowed("margin:0;")
-            >
-                <Button icon=IconName::Database>{"Queries"}</Button>
-                <Button icon=IconName::Function>{"Functions"}</Button>
-                <Button icon=IconName::Cog>{"Options"}</Button>
-            </ButtonGroup>
-        }
+#[function_component(Example)]
+pub fn example(props: &ExampleProps) -> Html {
+    html! {
+        <ButtonGroup
+            minimal={props.minimal}
+            fill={props.fill}
+            large={props.large}
+            vertical={props.vertical}
+            style={"margin:0;"}
+        >
+            <Button icon={IconName::Database}> {"Queries"}</Button>
+            <Button icon={IconName::Function}>{"Functions"}</Button>
+            <Button icon={IconName::Cog}>{"Options"}</Button>
+        </ButtonGroup>
     }
 }
