@@ -14,9 +14,9 @@ impl Component for CheckboxDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         CheckboxDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 disabled: false,
                 inline: false,
@@ -25,16 +25,12 @@ impl Component for CheckboxDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -43,18 +39,18 @@ impl Component for CheckboxDoc {
 
         html! {
             <div>
-            <H1 class=classes!("docs-title")>{"Checkbox"}</H1>
+            <H1 class={classes!("docs-title")}>{"Checkbox"}</H1>
             <SourceCodeUrl />
             <ExampleContainer
-                source=source
-                props=Some(html! {
+                source={source}
+                props={Some(html! {
                     <CheckboxProps
                         callback={self.callback.clone()}
-                        props=example_props.clone()
+                        example_props={example_props.clone()}
                     />
-                })
+                })}
             >
-                <Example with example_props />
+                <Example ..example_props />
             </ExampleContainer>
             </div>
         }
@@ -63,33 +59,33 @@ impl Component for CheckboxDoc {
 
 crate::build_example_prop_component! {
     CheckboxProps for ExampleProps =>
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         html! {
             <div>
                 <H5>{"Props"}</H5>
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         disabled: !props.disabled,
                         ..props
-                    })
-                    checked=self.props.disabled
-                    label=html!("Disabled")
+                    })}
+                    checked={ctx.props().example_props.disabled}
+                    label={html!("Disabled")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         inline: !props.inline,
                         ..props
-                    })
-                    checked=self.props.inline
-                    label=html!("Inline")
+                    })}
+                    checked={ctx.props().example_props.inline}
+                    label={html!("Inline")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         large: !props.large,
                         ..props
-                    })
-                    checked=self.props.large
-                    label=html!("Large")
+                    })}
+                    checked={ctx.props().example_props.large}
+                    label={html!("Large")}
                 />
             </div>
         }
