@@ -2,7 +2,6 @@ use yew::prelude::*;
 use yewprint::{Button, Collapse};
 
 pub struct Example {
-    link: ComponentLink<Self>,
     collapsed: bool,
 }
 
@@ -14,34 +13,27 @@ impl Component for Example {
     type Message = Msg;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        Example {
-            collapsed: true,
-            link,
-        }
+    fn create(_ctx: &Context<Self>) -> Self {
+        Example { collapsed: true }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         match msg {
             Msg::ToggleCollapse => self.collapsed ^= true,
         }
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         let logs = include_str!("example.log");
 
         html! {
             <div>
-                <Button onclick=self.link.callback(|_| Msg::ToggleCollapse)>
+                <Button onclick={ctx.link().callback(|_| Msg::ToggleCollapse)}>
                     {"Toggle collapse"}
                 </Button>
                 <Collapse
-                    is_open=!self.collapsed
+                    is_open={!self.collapsed}
                     keep_children_mounted=true
                 >
                     <pre class="bp3-code-block">
