@@ -3,8 +3,6 @@ use yewprint::{Button, Collapse, IconName, Intent};
 
 pub struct ExampleContainer {
     collapsed: bool,
-    props: ExampleContainerProps,
-    link: ComponentLink<Self>,
 }
 
 pub enum Msg {
@@ -23,41 +21,28 @@ impl Component for ExampleContainer {
     type Message = Msg;
     type Properties = ExampleContainerProps;
 
-    fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
-        ExampleContainer {
-            collapsed: true,
-            props,
-            link,
-        }
+    fn create(_ctx: &Context<Self>) -> Self {
+        ExampleContainer { collapsed: true }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         match msg {
             Msg::ToggleSource => self.collapsed ^= true,
         }
         true
     }
 
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         html! {
-            <div class=classes!("docs-example-wrapper")>
-                <div class=classes!("docs-example-frame", "docs-example-frame-row")>
-                    <div class=classes!("docs-example")>
-                        {self.props.children.clone()}
+            <div class={classes!("docs-example-wrapper")}>
+                <div class={classes!("docs-example-frame", "docs-example-frame-row")}>
+                    <div class={classes!("docs-example")}>
+                        {ctx.props().children.clone()}
                     </div>
                     {
-                        if let Some(props) = self.props.props.clone() {
+                        if let Some(props) = ctx.props().props.clone() {
                             html! {
-                                <div class=classes!("docs-example-options")>
+                                <div class={classes!("docs-example-options")}>
                                     {props}
                                 </div>
                             }
@@ -66,21 +51,21 @@ impl Component for ExampleContainer {
                         }
                     }
                 </div>
-                <div class=classes!("docs-source")>
+                <div class={classes!("docs-source")}>
                     <Button
-                        icon=IconName::Code
-                        fill={true}
-                        intent={Intent::Primary}
-                        minimal={true}
-                        onclick=self.link.callback(|_| Msg::ToggleSource)
+                        icon={IconName::Code}
+                        fill={{true}}
+                        intent={{Intent::Primary}}
+                        minimal={{true}}
+                        onclick={ctx.link().callback(|_| Msg::ToggleSource)}
                     >
                         {"View source"}
                     </Button>
                     <Collapse
-                        is_open=!self.collapsed
+                        is_open={!self.collapsed}
                         keep_children_mounted=true
                     >
-                        {self.props.source.clone()}
+                        {ctx.props().source.clone()}
                     </Collapse>
                 </div>
             </div>
@@ -88,47 +73,40 @@ impl Component for ExampleContainer {
     }
 }
 
+/// The macro generates the component that will be used to render the editable
+/// properties of an associated example.
 #[macro_export]
 macro_rules! build_example_prop_component {
     ($name:ident for $prop_component:ty => $($view:tt)*) => {
         #[derive(Clone, PartialEq, Properties)]
         pub struct $name {
             callback: Callback<$prop_component>,
-            props: $prop_component,
+            example_props: $prop_component
         }
 
         impl Component for $name {
             type Message = ();
             type Properties = Self;
 
-            fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-                props
-            }
-
-            fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-                true
-            }
-
-            fn change(&mut self, props: Self::Properties) -> ShouldRender {
-                if self.props != props.props || self.callback != props.callback {
-                    self.props = props.props;
-                    self.callback = props.callback;
-                    true
-                } else {
-                    false
-                }
+            fn create(ctx: &Context<Self>) -> Self {
+                ctx.props().clone()
             }
 
             $($view)*
         }
 
         impl $name {
+            /// Propagate the prop changes to the parent so the example is
+            /// re-rendered.
             fn update_props<T>(
                 &self,
+                props: &Self,
                 updater: impl Fn($prop_component, T) -> $prop_component + 'static,
             ) -> Callback<T> {
-                let props = self.props.clone();
-                self.callback.clone().reform(move |event| updater(props.clone(), event))
+                let example_props = props.example_props.clone();
+                self.callback
+                    .clone()
+                    .reform(move |event| updater(example_props.clone(), event))
             }
         }
     };
