@@ -14,9 +14,9 @@ impl Component for HtmlSelectDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         HtmlSelectDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 minimal: false,
                 fill: false,
@@ -26,16 +26,12 @@ impl Component for HtmlSelectDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -44,19 +40,19 @@ impl Component for HtmlSelectDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"HtmlSelect"}</H1>
+                <H1 class={classes!("docs-title")}>{"HtmlSelect"}</H1>
                 <SourceCodeUrl />
                 <div>
                     <ExampleContainer
-                        source=source
-                        props=Some(html! {
+                        source={source}
+                        props={Some(html! {
                             <HtmlSelectProps
                                 callback={self.callback.clone()}
-                                props=example_props.clone()
+                                example_props={example_props.clone()}
                             />
-                        })
+                        })}
                     >
-                        <Example with example_props />
+                        <Example ..example_props />
                     </ExampleContainer>
                 </div>
             </div>
@@ -66,41 +62,41 @@ impl Component for HtmlSelectDoc {
 
 crate::build_example_prop_component! {
     HtmlSelectProps for ExampleProps =>
-        fn view(&self) -> Html {
+        fn view(&self, ctx: &Context<Self>) -> Html {
             html! {
                 <div>
                     <H5>{"Props"}</H5>
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         minimal: !props.minimal,
                         ..props
-                    })
-                    checked=self.props.minimal
-                    label=html!("Minimal")
+                    })}
+                    checked={ctx.props().example_props.minimal}
+                    label={html!("Minimal")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps{
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps{
                         fill: !props.fill,
                         ..props
-                    })
-                    checked=self.props.fill
-                    label=html!("Fill")
+                    })}
+                    checked={ctx.props().example_props.fill}
+                    label={html!("Fill")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         disabled: !props.disabled,
                         ..props
-                    })
-                    checked=self.props.disabled
-                    label=html!("Disabled")
+                    })}
+                    checked={ctx.props().example_props.disabled}
+                    label={html!("Disabled")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps{
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps{
                         large: !props.large,
                         ..props
-                    })
-                    checked=self.props.large
-                    label=html!("Large")
+                    })}
+                    checked={ctx.props().example_props.large}
+                    label={html!("Large")}
                 />
             </div>
             }
