@@ -1,11 +1,8 @@
 use crate::{DocMenu, Logo};
-use std::borrow::Cow;
 use yew::prelude::*;
 use yewprint::{Icon, IconName, Menu, MenuDivider, MenuItem};
 
-pub struct Example {
-    link: ComponentLink<Self>,
-}
+pub struct Example {}
 
 pub enum Msg {
     GoToMenu(DocMenu),
@@ -15,107 +12,106 @@ impl Component for Example {
     type Message = Msg;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        Self { link }
+    fn create(_ctx: &Context<Self>) -> Self {
+        Self {}
     }
 
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, _msg: Self::Message) -> bool {
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
+    fn view(&self, ctx: &Context<Self>) -> Html {
+        let custom_icon = html! {
+            <Logo class={classes!("custom-icon")} />
+        };
 
-    fn view(&self) -> Html {
+        let share_icon = html! {
+            <Icon icon={IconName::Share} />
+        };
+
         html! {
             <>
                 <Menu>
                     <MenuItem
                         text={html!("Custom SVG icon")}
-                        icon_html=html! {
-                            <Logo class=classes!("custom-icon") />
-                        }
+                        icon_html={custom_icon}
                     />
                     <MenuDivider />
                     <MenuItem
-                        icon=IconName::NewTextBox
-                        text={html!{"New text box"}}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        icon={IconName::NewTextBox}
+                        text={html!("New text box")}
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::NewObject
-                        text={html!{"New object"}}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        icon={IconName::NewObject}
+                        text={html!("New object")}
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::NewLink
-                        text={html!{"New link"}}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        icon={IconName::NewLink}
+                        text={html!("New link")}
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuDivider />
                     <MenuItem
-                        icon=IconName::Cog
-                        text={html!{"Settings"}}
-                        label=html! {
-                            <Icon icon=IconName::Share />
-                        }
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        icon={IconName::Cog}
+                        text={html!("Settings")}
+                        label={share_icon}
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                 </Menu>
                 <Menu>
                     <MenuDivider title={html!("Edit")} />
                     <MenuItem
-                        icon=IconName::Cut
+                        icon={IconName::Cut}
                         text={html!("Cut")}
                         label={html!("Ctrl+X")}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::Duplicate
+                        icon={IconName::Duplicate}
                         text={html!("Copy")}
                         label={html!("Ctrl+C")}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::Clipboard
+                        icon={IconName::Clipboard}
                         text={html!("Paste")}
                         label={html!("Ctrl+V")}
                         disabled=true
                     />
                     <MenuDivider title={html!("Text")} />
                     <MenuItem
-                        icon=IconName::AlignLeft
+                        icon={IconName::AlignLeft}
                         text={html!("Alignment")}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::Style
+                        icon={IconName::Style}
                         text={html!("Style")}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        href={"#menu"}
+                        onclick={ctx.link()
+                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                     <MenuItem
-                        icon=IconName::Asterisk
+                        icon={IconName::Asterisk}
                         text={html!("Miscellaneous")}
-                        href=Cow::Borrowed("#menu")
-                        onclick=self.link
-                            .callback(|_| Msg::GoToMenu(DocMenu::Menu))
+                        href={"#menu"}
+                        onclick={ctx.link().callback(|_| Msg::GoToMenu(DocMenu::Menu))}
                     />
                 </Menu>
             </>
