@@ -14,9 +14,9 @@ impl Component for NumericInputDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         NumericInputDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 fill: false,
                 disabled: false,
@@ -28,16 +28,12 @@ impl Component for NumericInputDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -46,19 +42,19 @@ impl Component for NumericInputDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"NumericInput"}</H1>
+                <H1 class={classes!("docs-title")}>{"NumericInput"}</H1>
                 <SourceCodeUrl />
                 <div>
                     <ExampleContainer
-                        source=source
-                        props=Some(html! {
+                        source={source}
+                        props={Some(html! {
                             <NumericInputProps
                                 callback={self.callback.clone()}
-                                props=example_props.clone()
+                                example_props={example_props.clone()}
                             />
-                        })
+                        })}
                     >
-                        <Example with example_props />
+                        <Example ..example_props />
                     </ExampleContainer>
                 </div>
             </div>
@@ -68,57 +64,57 @@ impl Component for NumericInputDoc {
 
 crate::build_example_prop_component! {
     NumericInputProps for ExampleProps =>
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         html! {
             <div>
                 <H5>{"Props"}</H5>
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         fill: !props.fill,
                         ..props
-                    })
-                    checked=self.props.fill
-                    label=html!("Fill")
+                    })}
+                    checked={ctx.props().example_props.fill}
+                    label={html!("Fill")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         disabled: !props.disabled,
                         ..props
-                    })
-                    checked=self.props.disabled
-                    label=html!("Disabled")
+                    })}
+                    checked={ctx.props().example_props.disabled}
+                    label={html!("Disabled")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         large: !props.large,
                         ..props
-                    })
-                    checked=self.props.large
-                    label=html!("Large")
+                    })}
+                    checked={ctx.props().example_props.large}
+                    label={html!("Large")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         disable_buttons: !props.disable_buttons,
                         ..props
-                    })
-                    checked=self.props.disable_buttons
-                    label=html!("Disable buttons")
+                    })}
+                    checked={ctx.props().example_props.disable_buttons}
+                    label={html!("Disable buttons")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         buttons_on_the_left: !props.buttons_on_the_left,
                         ..props
-                    })
-                    checked=self.props.buttons_on_the_left
-                    label=html!("Buttons on the left")
+                    })}
+                    checked={ctx.props().example_props.buttons_on_the_left}
+                    label={html!("Buttons on the left")}
                 />
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         left_icon: !props.left_icon,
                         ..props
-                    })
-                    checked=self.props.left_icon
-                    label=html!("Left icon")
+                    })}
+                    checked={ctx.props().example_props.left_icon}
+                    label={html!("Left icon")}
                 />
             </div>
         }
