@@ -1,10 +1,6 @@
 use yew::prelude::*;
 use yewprint::{Intent, ProgressBar};
 
-pub struct Example {
-    props: ExampleProps,
-}
-
 #[derive(Clone, PartialEq, Properties)]
 pub struct ExampleProps {
     pub intent: Option<Intent>,
@@ -12,32 +8,14 @@ pub struct ExampleProps {
     pub stripes: bool,
 }
 
-impl Component for Example {
-    type Message = ();
-    type Properties = ExampleProps;
-
-    fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-        Example { props }
-    }
-    fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-        true
-    }
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-    fn view(&self) -> Html {
-        html! {
-            <ProgressBar
-                animate=self.props.animate
-                stripes=self.props.stripes
-                intent=self.props.intent
-                value=0.3
-            />
-        }
+#[function_component(Example)]
+pub fn example(props: &ExampleProps) -> Html {
+    html! {
+        <ProgressBar
+            animate={props.animate}
+            stripes={props.stripes}
+            intent={props.intent}
+            value=0.3
+        />
     }
 }
