@@ -2,8 +2,6 @@ use yew::prelude::*;
 use yewprint::{Label, Radio, RadioGroup};
 
 pub struct Example {
-    props: ExampleProps,
-    link: ComponentLink<Self>,
     selected_value: Lunch,
 }
 
@@ -22,15 +20,13 @@ impl Component for Example {
     type Message = Msg;
     type Properties = ExampleProps;
 
-    fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(_ctx: &Context<Self>) -> Self {
         Example {
-            props,
             selected_value: Lunch::Salad,
-            link,
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         match msg {
             Msg::ValueUpdate(value) => {
                 self.selected_value = value;
@@ -39,51 +35,42 @@ impl Component for Example {
         }
     }
 
-    fn change(&mut self, props: Self::Properties) -> ShouldRender {
-        if self.props != props {
-            self.props = props;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         html! {
             <>
                 <div>
                     <Radio
-                        label=html!("Blue pill")
-                        inline=self.props.inline
-                        disabled=self.props.disabled
-                        large=self.props.large
-                        name="group".to_string()
+                        label={html!("Blue pill")}
+                        inline={ctx.props().inline}
+                        disabled={ctx.props().disabled}
+                        large={ctx.props().large}
+                        name={"group".to_string()}
                     />
                     <Radio
-                        label=html!("Red pill")
-                        inline=self.props.inline
-                        disabled=self.props.disabled
-                        large=self.props.large
-                        name="group".to_string()
+                        label={html!("Red pill")}
+                        inline={ctx.props().inline}
+                        disabled={ctx.props().disabled}
+                        large={ctx.props().large}
+                        name={"group".to_string()}
                     />
                 </div>
                 <div>
                     <RadioGroup<Lunch>
-                        label=Some(html!(
+                        label={Some(html!(
                             <Label>
                                 {"Determine Lunch"}
                             </Label>
-                        ))
-                        options=vec![
+                        ))}
+                        options={vec![
                             (Lunch::Soup, "Soup".to_string()),
                             (Lunch::Salad, "Salad".to_string()),
                             (Lunch::Sandwich, "Sandwich".to_string()),
-                        ]
-                        value=self.selected_value
-                        onchange=self.link.callback(|v| Msg::ValueUpdate(v))
-                        inline=self.props.inline
-                        disabled=self.props.disabled
-                        large=self.props.large
+                        ]}
+                        value={self.selected_value}
+                        onchange={ctx.link().callback(|v| Msg::ValueUpdate(v))}
+                        inline={ctx.props().inline}
+                        disabled={ctx.props().disabled}
+                        large={ctx.props().large}
                     />
                 </div>
             </>
