@@ -14,9 +14,9 @@ impl Component for SliderDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         SliderDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 vertical: false,
                 intent: None,
@@ -24,16 +24,12 @@ impl Component for SliderDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -42,21 +38,21 @@ impl Component for SliderDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"Slider"}</H1>
+                <H1 class={classes!("docs-title")}>{"Slider"}</H1>
                 <SourceCodeUrl />
                 <ExampleContainer
-                    source=source
-                    props=Some(html! {
+                    source={source}
+                    props={Some(html! {
                         <SliderProps
                             callback={self.callback.clone()}
-                            props=example_props.clone()
+                            example_props={example_props.clone()}
                         />
-                    })
+                    })}
                 >
-                    <Example with example_props />
+                    <Example ..example_props />
                 </ExampleContainer>
                 <H2>{"Edge Cases"}</H2>
-                <div class=classes!("bp3-running-text")>
+                <div class={classes!("bp3-running-text")}>
                     <ul>
                         <li>
                             <p>
@@ -108,17 +104,17 @@ impl Component for SliderDoc {
 
 crate::build_example_prop_component! {
     SliderProps for ExampleProps =>
-    fn view(&self) -> Html {
+    fn view(&self, ctx: &Context<Self>) -> Html {
         html! {
             <div>
                 <H5>{"Props"}</H5>
                 <Switch
-                    onclick=self.update_props(|props, _| ExampleProps {
+                    onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                         vertical: !props.vertical,
                         ..props
-                    })
-                    checked=self.props.vertical
-                    label=html!("Vertical")
+                    })}
+                    checked={ctx.props().example_props.vertical}
+                    label={html!("Vertical")}
                     disabled=true
                 />
                 <p>{"Select intent:"}</p>
@@ -130,10 +126,10 @@ crate::build_example_prop_component! {
                         (Some(Intent::Warning), "Warning".to_string()),
                         (Some(Intent::Danger), "Danger".to_string()),
                     ]}
-                    onchange=self.update_props(|props, intent| ExampleProps {
+                    onchange={self.update_props(ctx.props(), |props, intent| ExampleProps {
                         intent,
                         ..props
-                    })
+                    })}
                 />
             </div>
         }
