@@ -14,9 +14,9 @@ impl Component for TabsDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         TabsDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 animate: true,
                 vertical: false,
@@ -24,16 +24,12 @@ impl Component for TabsDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -42,19 +38,19 @@ impl Component for TabsDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"Tabs"}</H1>
+                <H1 class={classes!("docs-title")}>{"Tabs"}</H1>
                 <SourceCodeUrl />
                 <div>
                     <ExampleContainer
-                        source=source
-                        props=Some(html! {
+                        source={source}
+                        props={Some(html! {
                             <TabsProps
                                 callback={self.callback.clone()}
-                                props=example_props.clone()
+                                example_props={example_props.clone()}
                             />
-                        })
+                        })}
                     >
-                        <Example with example_props />
+                        <Example ..example_props />
                     </ExampleContainer>
                 </div>
             </div>
@@ -64,25 +60,25 @@ impl Component for TabsDoc {
 
 crate::build_example_prop_component! {
     TabsProps for ExampleProps =>
-        fn view(&self) -> Html {
+        fn view(&self, ctx: &Context<Self>) -> Html {
             html! {
                 <div>
                     <H5>{"Props"}</H5>
                     <Switch
-                        onclick=self.update_props(|props, _| ExampleProps {
+                        onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                             animate: !props.animate,
                             ..props
-                        })
-                        checked=self.props.animate
-                        label=html!("Animate indicator")
+                        })}
+                        checked={ctx.props().example_props.animate}
+                        label={html!("Animate indicator")}
                     />
                     <Switch
-                        onclick=self.update_props(|props, _| ExampleProps {
+                        onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                             vertical: !props.vertical,
                             ..props
-                        })
-                        checked=self.props.vertical
-                        label=html!("Use vertical tabs")
+                        })}
+                        checked={ctx.props().example_props.vertical}
+                        label={html!("Use vertical tabs")}
                     />
                 </div>
             }
