@@ -28,9 +28,9 @@ impl Component for TagDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_props: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         TagDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 initial_tags: initial_tags(),
                 active: false,
@@ -49,16 +49,12 @@ impl Component for TagDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -67,18 +63,18 @@ impl Component for TagDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"Tag"}</H1>
+                <H1 class={classes!("docs-title")}>{"Tag"}</H1>
                 <SourceCodeUrl />
                 <ExampleContainer
-                    source=source
-                    props=Some(html!{
+                    source={source}
+                    props={Some(html!{
                         <TagProps
                             callback={self.callback.clone()}
-                            props=example_props.clone()
+                            example_props={example_props.clone()}
                         />
-                    })
+                    })}
                 >
-                    <Example with example_props/>
+                    <Example ..example_props/>
                 </ExampleContainer>
             </div>
         }
@@ -87,90 +83,92 @@ impl Component for TagDoc {
 
 crate::build_example_prop_component! {
     TagProps for ExampleProps =>
-        fn view(&self) -> Html {
+        fn view(&self, ctx: &Context<Self>) -> Html {
+            let props = ctx.props();
+
             html! {
                 <div>
                     <H5>{"Props"}</H5>
                     <div>
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 active: !props.active,
                                 ..props
-                            })
-                            checked=self.props.active
-                            label=html!("Active")
+                            })}
+                            checked={props.example_props.active}
+                            label={html!("Active")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 fill: !props.fill,
                                 ..props
-                            })
-                            checked=self.props.fill
-                            label=html!("Fill")
+                            })}
+                            checked={props.example_props.fill}
+                            label={html!("Fill")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 interactive: !props.interactive,
                                 ..props
-                            })
-                            checked=self.props.interactive
-                            label=html!("Interactive")
+                            })}
+                            checked={props.example_props.interactive}
+                            label={html!("Interactive")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 large: !props.large,
                                 ..props
-                            })
-                            checked=self.props.large
-                            label=html!("Large")
+                            })}
+                            checked={props.example_props.large}
+                            label={html!("Large")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 minimal: !props.minimal,
                                 ..props
-                            })
-                            checked=self.props.minimal
-                            label=html!("Minimal")
+                            })}
+                            checked={props.example_props.minimal}
+                            label={html!("Minimal")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 multiline: !props.multiline,
                                 ..props
-                            })
-                            checked=self.props.multiline
-                            label=html!("Multiline")
+                            })}
+                            checked={props.example_props.multiline}
+                            label={html!("Multiline")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 round: !props.round,
                                 ..props
-                            })
-                            checked=self.props.round
-                            label=html!("Round")
+                            })}
+                            checked={props.example_props.round}
+                            label={html!("Round")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 removable: !props.removable,
                                 ..props
-                            })
-                            checked=self.props.removable
-                            label=html!("Removable")
+                            })}
+                            checked={props.example_props.removable}
+                            label={html!("Removable")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 icon: !props.icon,
                                 ..props
-                            })
-                            checked=self.props.icon
-                            label=html!("Icon")
+                            })}
+                            checked={props.example_props.icon}
+                            label={html!("Icon")}
                         />
                         <Switch
-                            onclick=self.update_props(|props, _| ExampleProps {
+                            onclick={self.update_props(props, |props, _| ExampleProps {
                                 right_icon: !props.right_icon,
                                 ..props
-                            })
-                            checked=self.props.right_icon
-                            label=html!("Right icon")
+                            })}
+                            checked={props.example_props.right_icon}
+                            label={html!("Right icon")}
                         />
                         <p>{"Select intent:"}</p>
                         <ButtonGroup
@@ -184,17 +182,17 @@ crate::build_example_prop_component! {
                                     (Some(Intent::Warning), "Warning".to_string()),
                                     (Some(Intent::Danger), "Danger".to_string()),
                                 ]}
-                                onchange=self.update_props(|props, intent| ExampleProps {
+                                onchange={self.update_props(props, |props, intent| ExampleProps {
                                     intent,
                                     ..props
-                                })
+                                })}
                             />
                             <Button
-                                icon=IconName::Refresh
-                                onclick=self.update_props(|props, _| ExampleProps {
+                                icon={IconName::Refresh}
+                                onclick={self.update_props(props, |props, _| ExampleProps {
                                     reset_tags: props.reset_tags + 1,
-                                    ..props
-                                })
+                                    ..props.clone()
+                                })}
                             >
                                 {"Reset tags"}
                             </Button>
