@@ -2,6 +2,7 @@ mod example;
 
 use crate::ExampleContainer;
 use example::*;
+use web_sys::HtmlInputElement;
 use yew::prelude::*;
 use yewprint::{Switch, H1, H5};
 
@@ -14,9 +15,9 @@ impl Component for TextDoc {
     type Message = ExampleProps;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
+    fn create(ctx: &Context<Self>) -> Self {
         TextDoc {
-            callback: link.callback(|x| x),
+            callback: ctx.link().callback(|x| x),
             state: ExampleProps {
                 ellipsize: false,
                 text: String::from("Hello, world!"),
@@ -24,16 +25,12 @@ impl Component for TextDoc {
         }
     }
 
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         self.state = msg;
         true
     }
 
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        true
-    }
-
-    fn view(&self) -> Html {
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         let example_props = self.state.clone();
         let source = crate::include_raw_html!(
             concat!(env!("OUT_DIR"), "/", file!(), ".html"),
@@ -42,19 +39,19 @@ impl Component for TextDoc {
 
         html! {
             <div>
-                <H1 class=classes!("docs-title")>{"Text"}</H1>
+                <H1 class={classes!("docs-title")}>{"Text"}</H1>
                 <SourceCodeUrl />
                 <div>
                     <ExampleContainer
-                        source=source
-                        props=Some(html! {
+                        source={source}
+                        props={Some(html! {
                             <TextProps
                                 callback={self.callback.clone()}
-                                props=example_props.clone()
+                                example_props={example_props.clone()}
                             />
-                        })
+                        })}
                     >
-                        <Example with example_props />
+                        <Example ..example_props />
                     </ExampleContainer>
                 </div>
             </div>
@@ -64,37 +61,36 @@ impl Component for TextDoc {
 
 crate::build_example_prop_component! {
     TextProps for ExampleProps =>
-        fn view(&self) -> Html {
+        fn view(&self, ctx: &Context<Self>) -> Html {
             html! {
                 <div>
                     <H5>{"Props"}</H5>
                     <Switch
-                        onclick=self.update_props(|props, _| ExampleProps {
+                        onclick={self.update_props(ctx.props(), |props, _| ExampleProps {
                             ellipsize: !props.ellipsize,
                             ..props
-                        })
-                        checked=self.props.ellipsize
-                        label=html!("Ellipsize")
+                        })}
+                        checked={ctx.props().example_props.ellipsize}
+                        label={html!("Ellipsize")}
                     />
                     <input
                         class="bp3-input"
-                        onchange=self.update_props(|props, e|
-                            match e {
-                                ChangeData::Value(text) => {
+                        onchange={self.update_props(ctx.props(), |props, e: Event| {
+                                if let Some(input) = e.target_dyn_into::<HtmlInputElement>() {
                                     ExampleProps {
-                                        text,
+                                        text: input.value(),
                                         ..props
                                     }
-                                },
-                                _ => {
+                                } else {
                                     ExampleProps {
                                         text: "Hello, world!".to_string(),
                                         ..props
                                     }
                                 }
-                        })
+                            }
+                        )}
                         type="text"
-                        value={self.props.text.clone()}
+                        value={ctx.props().example_props.text.clone()}
                     />
                 </div>
             }
